+++ examples/resources/iosxe_device_tracking/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_device_tracking.example ""