+++ examples/resources/iosxe_interface_port_channel_subinterface/resource.tf
@@ -1,5 +1,5 @@
 resource "iosxe_interface_port_channel_subinterface" "example" {
-  name                        = "10.666"
+  name                        = "20.666"
   encapsulation_dot1q_vlan_id = 666
   description                 = "My Interface Description"
   shutdown                    = false
@@ -40,4 +40,10 @@ resource "iosxe_interface_port_channel_subinterface" "example" {
   ]
   arp_timeout     = 2147
   ip_igmp_version = 3
+  ip_flow_monitors = [
+    {
+      name      = "MON2"
+      direction = "input"
+    }
+  ]
 }