+++ internal/provider/resource_iosxe_interface_port_channel_subinterface.go
@@ -340,6 +340,25 @@ func (r *InterfacePortChannelSubinterfaceResource) Schema(ctx context.Context, r
 					},
 				},
 			},
+			"ipv6_flow_monitors": schema.ListNestedAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Apply a Flow Monitor").String,
+				Optional:            true,
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"name": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("User defined").String,
+							Required:            true,
+						},
+						"direction": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("").AddStringEnumDescription("input", "output").String,
+							Required:            true,
+							Validators: []validator.String{
+								stringvalidator.OneOf("input", "output"),
+							},
+						},
+					},
+				},
+			},
 			"arp_timeout": schema.Int64Attribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Set ARP cache timeout").AddIntegerRangeDescription(0, 2147483).String,
 				Optional:            true,
@@ -376,6 +395,25 @@ func (r *InterfacePortChannelSubinterfaceResource) Schema(ctx context.Context, r
 			"ip_nat_outside": schema.BoolAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Outside interface for address translation").String,
 				Optional:            true,
+			},
+			"ip_flow_monitors": schema.ListNestedAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Apply a Flow Monitor").String,
+				Optional:            true,
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"name": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("User defined").String,
+							Required:            true,
+						},
+						"direction": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("").AddStringEnumDescription("input", "output").String,
+							Required:            true,
+							Validators: []validator.String{
+								stringvalidator.OneOf("input", "output"),
+							},
+						},
+					},
+				},
 			},
 		},
 	}